# api-gateway/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: Learning Platform API
  description: API для управления образовательной платформой
  version: '1.0'
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
...
  /auth/token/refresh:
    post:
      tags:
        - Auth Controller
      summary: Обновление токена
      operationId: refreshToken
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/JwtAuthResponse'
  /auth/reg:
    post:
      tags:
        - Auth Controller
      summary: Регистрация пользователя
      description: Позволяет зарегистрировать пользователя с ролью USER
      operationId: registration
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegRequest'
        required: true
      responses:
        '200':
          description: Регистрация успешна
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/JwtAuthResponse'
        '400':
          description: Пользователь уже существует
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/JwtAuthResponse'
  /auth/login:
    post:
      tags:
        - Auth Controller
      summary: Авторизация пользователя
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/JwtAuthResponse'
  /access/module/get-access:
    post:
      tags:
        - Access Controller
      summary: Выдать доступ к модулю
      description: Позволяет выдать доступ пользователю к указанному модулю курса
      operationId: giveModuleAccess
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModuleAccessRequest'
        required: true
      responses:
        '200':
          description: Доступ успешно выдан
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                type: object
        '403':
          description: Доступ запрещен
          content:
            '*/*':
              schema:
                type: object
        '404':
          description: Модуль или пользователь не найден
          content:
            '*/*':
              schema:
                type: object
  /access/module/delete-access:
    post:
      tags:
        - Access Controller
      summary: Отозвать доступ к модулю
      description: Позволяет отозвать доступ пользователя к указанному модулю курса
      operationId: revokeModuleAccess
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModuleAccessRequest'
        required: true
      responses:
        '200':
          description: Доступ успешно отозван
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                type: object
        '403':
          description: Доступ запрещен
          content:
            '*/*':
              schema:
                type: object
        '404':
          description: Доступ не найден
          content:
            '*/*':
              schema:
                type: object
  /access/course/get-access:
    post:
      tags:
        - Access Controller
      summary: Выдать доступ к курсу
      description: Позволяет выдать доступ пользователю к указанному курсу
      operationId: giveCourseAccess
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseAccessRequest'
        required: true
      responses:
        '200':
          description: Доступ успешно выдан
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                type: object
        '403':
          description: Доступ запрещен
          content:
            '*/*':
              schema:
                type: object
        '404':
          description: Курс или пользователь не найден
          content:
            '*/*':
              schema:
                type: object
  /access/course/delete-access:
    post:
      tags:
        - Access Controller
      summary: Отозвать доступ к курсу
      description: Позволяет отозвать доступ пользователя к указанному курсу
      operationId: revokeCourseAccess
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseAccessRequest'
        required: true
      responses:
        '200':
          description: Доступ успешно отозван
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                type: object
        '403':
          description: Доступ запрещен
          content:
            '*/*':
              schema:
                type: object
        '404':
          description: Доступ не найден
          content:
            '*/*':
              schema:
                type: object
...
components:
  schemas:
...
    JwtAuthResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT токен доступа
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refreshToken:
          type: string
          description: JWT токен обновления
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        role:
          type: string
          description: Роль пользователя в системе
          example: USER
          enum:
            - USER
            - ADMIN
            - MENTOR
            - USER
            - MENTOR
            - ADMIN
      description: Ответ с токенами доступа и ролью пользователя
    RegRequest:
      required:
        - firstName
        - lastName
      type: object
      properties:
        username:
          maxLength: 50
          minLength: 5
          type: string
          description: Email пользователя
          example: pipa@popa.com
        password:
          maxLength: 255
          minLength: 6
          pattern: \S+
          type: string
          description: Пароль
          example: my_1secret1_password
        tgNickname:
          maxLength: 32
          minLength: 5
          pattern: ^@[a-zA-Z0-9_]+
          type: string
          description: Никнейм в Telegram
          example: '@john_doe'
        firstName:
          type: string
          description: Имя пользователя
          example: Иван
        lastName:
          type: string
          description: Фамилия пользователя
          example: Иванов
        confirmPassword:
          type: string
          description: Подтверждение пароля
          example: my_1secret1_password
      description: Запрос на регистрацию нового пользователя
    AuthRequest:
      type: object
      properties:
        username:
          maxLength: 50
          minLength: 5
          type: string
          description: Email пользователя
          example: pipa@popa.com
        password:
          maxLength: 255
          minLength: 6
          pattern: \S+
          type: string
          description: Пароль
          example: my_1secret1_password
      description: Запрос на авторизацию
    ModuleAccessRequest:
      required:
        - courseId
        - userId
      type: object
      properties:
        userId:
          type: integer
          description: ID пользователя
          format: int64
          example: 123
        courseId:
          type: integer
          description: ID курса
          format: int64
          example: 456
        moduleId:
          type: integer
          description: ID модуля
          format: int64
          example: 789
      description: Запрос на управление доступом к модулю курса
    CourseAccessRequest:
      required:
        - courseId
        - userId
      type: object
      properties:
        userId:
          type: integer
          description: ID пользователя
          format: int64
          example: 123
        courseId:
          type: integer
          description: ID курса
          format: int64
          example: 456
      description: Запрос на управление доступом к курсу
...